Move constants into separate package
